+++ CHANGELOG.md
@@ -7,9 +7,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
-## [0.5.0] - 2021-04-16
-
 ### Added
+- Implement magic scaling factor for large filesystems similar to original ruby plugin.
 - Added --human-readable option to support ignoring human-readable option like df unix/linux command
 
 ## [0.4.2] - 2021-03-31